write_to_file -> save_castle
and various magic numbers or macros

+++ src/config.h
@@ -1,4 +1,4 @@
-#define DEV_undefined -1
+#define DEV_undefined (byte)  -1
 #define DEV_keyboard 0
 #define DEV_joystick 1
 extern char controller;
